Move Lifetime to its own file
