+++ glean/src/core/metrics/index.ts
@@ -6,18 +6,7 @@ import { isUndefined, JSONValue } from "../utils.js";
 import Glean from "../glean.js";
 import LabeledMetricType from "./types/labeled.js";
 import { Metric } from "./metric.js";
-
-/**
- * An enum representing the possible metric lifetimes.
- */
-export const enum Lifetime {
-  // The metric is reset with each sent ping
-  Ping = "ping",
-  // The metric is reset on application restart
-  Application = "application",
-  // The metric is reset with each user profile
-  User = "user",
-}
+import { Lifetime } from "./lifetime.js";
 
 /**
  * The common set of data shared across all different metric types.
