+++ glean/tests/core/glean.spec.ts
@@ -8,13 +8,13 @@ import sinon from "sinon";
 import { CLIENT_INFO_STORAGE, DELETION_REQUEST_PING_NAME, KNOWN_CLIENT_ID } from "../../src/core/constants";
 import CoreEvents from "../../src/core/events";
 import Glean from "../../src/core/glean";
-import { Lifetime } from "../../src/core/metrics";
 import StringMetricType from "../../src/core/metrics/types/string";
 import CounterMetricType from "../../src/core/metrics/types/counter";
 import PingType from "../../src/core/pings";
 import { isObject, JSONObject } from "../../src/core/utils";
 import TestPlatform from "../../src/platform/qt";
 import Plugin from "../../src/plugins";
+import { Lifetime } from "../../src/core/metrics/lifetime";
 
 class MockPlugin extends Plugin<typeof CoreEvents["afterPingCollection"]> {
   constructor() {
