+++ glean/tests/core/metrics/datetime.spec.ts
@@ -7,8 +7,8 @@ import sinon from "sinon";
 
 import Glean from "../../../src/core/glean";
 import DatetimeMetricType, { DatetimeMetric } from "../../../src/core/metrics/types/datetime";
-import { Lifetime } from "../../../src/core/metrics";
 import TimeUnit from "../../../src/core/metrics/time_unit";
+import { Lifetime } from "../../../src/core/metrics/lifetime";
 
 const sandbox = sinon.createSandbox();
