+++ glean/tests/core/metrics/string.spec.ts
@@ -6,7 +6,7 @@ import assert from "assert";
 
 import Glean from "../../../src/core/glean";
 import StringMetricType, { MAX_LENGTH_VALUE } from "../../../src/core/metrics/types/string";
-import { Lifetime } from "../../../src/core/metrics";
+import { Lifetime } from "../../../src/core/metrics/lifetime";
  
 describe("StringMetric", function() {
   const testAppId = `gleanjs.test.${this.title}`;
